Changed temp range from "40 to 90" to "40 to 89"
The math in the examples (and in the problem's solution) would not be correct if the temp could be 90 degrees. (Because it would take 51 clicks to reach 40 degrees again, and -1 clicks from 40 would bring you to 90 degrees, not 89). Quickest fix is to change the range to "40 to 89 degrees", which I did here.