Update Strings exercises solution for 3
The third exercise in the Strings chapter provided a solution that did not pass the tests. The previous RST version passed the tests, so I took most of the changes from there. I updated it a little bit so that the spacing of the printed message of the solution matches exactly with the presented message in the problem description. I also updated the problem description to mention that the number of 'e' characters that were found should be returned since the tests expect that behavior.